LPS-148669 Move MailEngine getSession(account) method to MailServiceImpl
