+++ portal-kernel/src/com/liferay/mail/kernel/service/MailService.java
@@ -14,6 +14,7 @@
 
 package com.liferay.mail.kernel.service;
 
+import com.liferay.mail.kernel.model.Account;
 import com.liferay.mail.kernel.model.Filter;
 import com.liferay.mail.kernel.model.MailMessage;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -51,6 +52,8 @@ public interface MailService {
 	public void deleteUser(long companyId, long userId);
 
 	public Session getSession();
+
+	public Session getSession(Account account);
 
 	public void sendEmail(MailMessage mailMessage);
